apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-genai>=0.8.0
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
